Implemented user authentication and RESTful user auth API.
Creating an user is done through `POST /users {username: <username>, password: <password>` and then authenticating a user is done through `GET /auth_user?username=<username>&password=<password>` which returns result true or false depending on if the username and password match.

The passwords are hashed and salted to a key. The key and salt are stored in the database for later retrieval. The password is not stored for security (and hash and salt is up to recommended of sha256 100000 times and urand for salt).