Started the port of the back-end to Flask.
Commit is pretty large. The outcome of this commit is that you can create accounts with POST /user/<username>, create access tokens with PSOT /login, and access user data from GET /users/<username>.

The common directory is used for modules that are used in various resources. The resources directory is used for code that correspond to e.g. /users = resources.users.